docker_mount.py Added ability to umount based on mount point only
For stand alone use of this class, applications like atomic
need to be able to umount without having the image_tuple.  In
this case, they can now umount things based on only knowing the
where the thin_pool was mounted.  We then build the tuple from
that and send to cleanup()