Output something if not RHEL scan.py, reporter.py, image-scanner.py
This commit is a fix for:
    https://github.com/baude/image-scanner/issues/12

Which is basically that if an image isnt RHEL, nothing was
being output to the summary.  If that was the only image
that needed to be scanned, it makes it look like it failed.
Now users will know we inspected the image and didn't scan
it because it isnt rhel.