* Allow broader inputs (container names, ids, image names)
image-scanner -i now accepts various inputs as input:

    - image id
    - image name
    - image name:tag
    - container id
    - contaier name

The allcontainers and onlyactive command line switches were
also changes to scan the containers rather than only the
images associated with the container.

The summary outputs were also changed to reflect this.  Now
if you scan a container, the Image: label is changed to
Container: and we removed the Containers effected label.

This patch resolves:
     issue #1 https://github.com/baude/image-scanner/issues/1